feat: Retry when some of batch operations are failing
Also added a test for retry on batch operation.

This implementation is not so elegant as far as I'm considering. Batch
operation could possibly be retried even when result is `Ok`. So I had
to wrap it in a `Result<(), Result<RpBatch>>` type. And the `errors`
method on `BatchedResult` requires dynamic dispatching.