Add PHP bindings for OpenDAL library
A new subdirectory `bindings/php` has been created in order to have PHP bindings for the OpenDAL library. This includes new PHP tests, stdlib stubs for the PHP interface, and addition of composer configuration for dependency management. as well as the initial source code written in Rust to provide PHP bindings. The 'debug' function is used as a test to check the binding. The README has been added to guide users and developers in building and testing the PHP bindings.